fxes to directory structure
