+++ Entities/Shade.cs
