+++ Patch/GatePatcher.cs
