+++ Patch/HeroControllerPatcher.cs
