+++ Patch/ShinyItemStandPatcher.cs
