+++ Patch/VignettePatcher.cs
