+++ SmolKnight.csproj
@@ -97,25 +97,25 @@
     <ItemGroup>
       <Compile Include="SmolKnight.cs" />
       <Compile Include="Properties\AssemblyInfo.cs" />
-      <Compile Include="SmolKnight\Menu\ModMenu.cs" />
+      <Compile Include="Options\ModMenu.cs" />
+      <Compile Include="Options\Settings.cs" />
 
-      <Compile Include="SmolKnight\Entities\KnightController.cs" />  
-      <Compile Include="SmolKnight\Entities\Knight.cs" />  
-      <Compile Include="SmolKnight\Entities\Shade.cs" />  
+      <Compile Include="Entities\KnightController.cs" />  
+      <Compile Include="Entities\Knight.cs" />  
+      <Compile Include="Entities\Shade.cs" />  
 
-      <Compile Include="SmolKnight\Patch\ActionPatcher.cs" />        
-      <Compile Include="SmolKnight\Patch\GatePatcher.cs" />        
-      <Compile Include="SmolKnight\Patch\ShinyItemStandPatcher.cs" />
-      <Compile Include="SmolKnight\Patch\VignettePatcher.cs" />
-      <Compile Include="SmolKnight\Patch\PlayerDataPatcher.cs" />
-      <Compile Include="SmolKnight\Patch\HeroControllerPatcher.cs" />
-      <Compile Include="SmolKnight\Patch\ILHooks.cs" />
+      <Compile Include="Patch\ActionPatcher.cs" />        
+      <Compile Include="Patch\GatePatcher.cs" />        
+      <Compile Include="Patch\ShinyItemStandPatcher.cs" />
+      <Compile Include="Patch\VignettePatcher.cs" />
+      <Compile Include="Patch\PlayerDataPatcher.cs" />
+      <Compile Include="Patch\HeroControllerPatcher.cs" />
+      <Compile Include="Patch\ILHooks.cs" />
 
-      <Compile Include="SmolKnight\HKMP.cs" />
-      <Compile Include="SmolKnight\Utils.cs" />
-      <Compile Include="SmolKnight\SFX.cs" />
-      <Compile Include="SmolKnight\Size.cs" />
-      <Compile Include="SmolKnight\Settings.cs" />
+      <Compile Include="Helpers\HKMP.cs" />
+      <Compile Include="Helpers\Utils.cs" />
+      <Compile Include="Helpers\SFX.cs" />
+      <Compile Include="Helpers\Size.cs" />
     </ItemGroup>
     <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
